aic {locfit}	R Documentation

Compute Akaike's Information Criterion.

Description

The calling sequence for aic matches those for the locfit or locfit.raw functions. The fit is not returned; instead, the returned object contains Akaike's information criterion for the fit.

The definition of AIC used here is -2*log-likelihood + pen*(fitted d.f.). For quasi-likelihood, and local regression, this assumes the scale parameter is one. Other scale parameters can effectively be used by changing the penalty.

The AIC score is exact (up to numerical roundoff) if the ev="data" argument is provided. Otherwise, the residual sum-of-squares and degrees of freedom are computed using locfit's standard interpolation based approximations.

Usage

aic(x, ..., pen=2)

Arguments

x
model formula

...
other arguments to locfit

pen
penalty for the degrees of freedom term

See Also

locfit, locfit.raw, aicplot

[Package locfit version 1.5-9.1 Index]