aicplot {locfit}	R Documentation

Compute an AIC plot.

Description

The aicplot function loops through calls to the aic function (and hence to locfit), using a different smoothing parameter for each call. The returned structure contains the AIC statistic for each fit, and can be used to produce an AIC plot.

Usage

aicplot(..., alpha)

Arguments

...
arguments to the aic, locfit functions.

alpha
Matrix of smoothing parameters. The aicplot function loops through calls to aic, using each row of alpha as the smoothing parameter in turn. If alpha is provided as a vector, it will be converted to a one-column matrix, thus interpreting each component as a nearest neighbor smoothing parameter.

Value

An object with class "gcvplot", containing the smoothing parameters and AIC scores. The actual plot is produced using plot.gcvplot.

See Also

locfit, locfit.raw, gcv, aic, plot.gcvplot

Examples

data(morths)
plot(aicplot(deaths~age,weights=n,data=morths,family="binomial",
  alpha=seq(0.2,1.0,by=0.05)))

[Package locfit version 1.5-9.1 Index]