chemdiab {locfit}	R Documentation

Chemical Diabetes Dataset

Description

Numeric variables are rw, fpg, ga, ina and sspg. Classifier cc is the Diabetic type.

Usage

data(chemdiab)

Format

Data frame with five numeric measurements and categroical response.

Source

Reaven and Miller (1979).

References

Reaven, G. M. and Miller, R. G. (1979). An attempt to define the nature of chemical diabetes using a multidimensional analysis. Diabetologia 16, 17-24.

[Package locfit version 1.5-9.1 Index]