co2 {locfit}	R Documentation

Carbon Dioxide Dataset

Description

Monthly time series of carbon dioxide measurements at Mauna Loa, Hawaii from 1959 to 1990.

Usage

data(co2)

Format

Data frame with year, month and co2 variables.

Source

Boden, Sepanski and Stoss (1992).

References

Boden, Sepanski and Stoss (1992). Trends '91: A compedium of data on global change - Highlights. Carbon Dioxide Information Analysis Center, Oak Ridge National Laboratory.

[Package locfit version 1.5-9.1 Index]