cpar {locfit}	R Documentation

Conditionally parametric term for a Locfit model.

Description

A term entered in a locfit model formula using cpar will result in a fit that is conditionally parametric. Equivalent to lp(x,style="cpar").

This function is presently almost deprecated. Specifying a conditionally parametric fit as y~x1+cpar(x2) wil no longer work; instead, the model is specified as y~lp(x1,x2,style=c("n","cpar")).

Usage

cpar(x,...)

Arguments

x
numeric variable.

...
Other arguments to link{lp}().

See Also

locfit

Examples

data(ethanol, package="locfit")
# fit a conditionally parametric model
fit <- locfit(NOx ~ lp(E, C, style=c("n","cpar")), data=ethanol)
plot(fit)
# one way to force a parametric fit with locfit
fit <- locfit(NOx ~ cpar(E), data=ethanol)

[Package locfit version 1.5-9.1 Index]