crit {locfit}	R Documentation

Compute critical values for confidence intervals.

Description

Every "locfit" object contains a critical value object to be used in computing and ploting confidence intervals. By default, a 95% pointwise confidence level is used. To change the confidence level, the critical value object must be substituted using crit and crit<-.

Usage

crit(fit, const=c(0, 1), d=1, cov=0.95, rdf=0)
crit(fit) <- value

Arguments

fit
"locfit" object. This is optional; if a fit is provided, defaults for the other arguments are taken from the critical value currently stored on this fit, rather than the usual values above. crit(fit) with no other arguments will just return the current critical value.

const
Tube formula constants for simultaneous bands (the default, c(0,1), produces pointwise coverage). Usually this is generated by the kappa0 function and should not be provided by the user.

d
Dimension of the fit. Again, users shouldn't usually provide it.

cov
Coverage Probability for critical values.

rdf
Residual degrees of freedom. If non-zero, the critical values are based on the Student's t distribution. When rdf=0, the normal distribution is used.

value
Critical value object generated by crit or kappa0.

Value

Critical value object.

See Also

locfit, plot.locfit, kappa0, crit<-.

Examples

# compute and plot 99% confidence intervals, with local variance estimate.
data(ethanol)
fit <- locfit(NOx~E,data=ethanol)
crit(fit) <- crit(fit,cov=0.99)
plot(fit,band="local")

# compute and plot 99% simultaneous bands
crit(fit) <- kappa0(NOx~E,data=ethanol,cov=0.99)
plot(fit,band="local")

[Package locfit version 1.5-9.1 Index]