density.lf {locfit}	R Documentation

Density estimation using Locfit

Description

This function provides an interface to Locfit, in the syntax of (a now old version of) the S-Plus density function. This can reproduce density results, but allows additional locfit.raw arguments, such as the degree of fit, to be given.

It also works in double precision, whereas density only works in single precision.

Usage

density.lf(x, n = 50, window = "gaussian", width, from, to,
  cut = if(iwindow == 4.) 0.75 else 0.5,
  ev = lfgrid(mg = n, ll = from, ur = to),
  deg = 0, family = "density", link = "ident", ...)

Arguments

x
numeric vector of observations whose density is to be estimated.

n
number of evaluation points. Equivalent to the locfit.raw mg argument.

window
Window type to use for estimation. Equivalent to the locfit.raw kern argument. This includes all the density windows except cosine.

width
Window width. Following density, this is the full width; not the half-width usually used by Locfit and many other smoothers.

from
Lower limit for estimation domain.

to
Upper limit for estimation domain.

cut
Controls default expansion of the domain.

ev
Locfit evaluation structure – default lfgrid().

deg
Fitting degree – default 0 for kernel estimation.

family
Fitting family – default is "density".

link
Link function – default is the "identity".

...
Additional arguments to locfit.raw, with standard defaults.

Value

A list with components x (evaluation points) and y (estimated density).

See Also

density, locfit, locfit.raw

Examples

data(geyser)
density.lf(geyser, window="tria")
# the same result with density, except less precision.
density(geyser, window="tria")

[Package locfit version 1.5-9.1 Index]