gam.slist {locfit}	R Documentation

Vector of GAM special terms

Description

This vector adds "lf" to the default vector of special terms recognized by a gam() model formula. To ensure this is recognized, attach the Locfit library with library(locfit,first=T).

Format

Character vector.

See Also

lf, gam

[Package locfit version 1.5-9.12 Index]