gcv {locfit}	R Documentation

Compute generalized cross-validation statistic.

Description

The calling sequence for gcv matches those for the locfit or locfit.raw functions. The fit is not returned; instead, the returned object contains Wahba's generalized cross-validation score for the fit.

The GCV score is exact (up to numerical roundoff) if the ev="data" argument is provided. Otherwise, the residual sum-of-squares and degrees of freedom are computed using locfit's standard interpolation based approximations.

For likelihood models, GCV is computed uses the deviance in place of the residual sum of squares. This produces useful results but I do not know of any theory validating this extension.

Usage

gcv(x, ...)

Arguments

x, ...
Arguments passed on to locfit or locfit.raw.

See Also

locfit, locfit.raw, gcvplot

[Package locfit version 1.5-9.1 Index]