geyser {locfit}	R Documentation

Old Faithful Geyser Dataset

Description

The durations of 107 eruptions of the Old Faithful Geyser.

Usage

data(geyser)

Format

A numeric vector of length 107.

Source

Scott (1992). Note that several different Old Faithful Geyser datasets (including the faithful dataset in R's base library) have been used in various places in the statistics literature. The version provided here has been used in density estimation and bandwidth selection work.

References

Scott, D. W. (1992). Multivariate Density Estimation: Theory, Practice and Visualization. Wiley.

[Package locfit version 1.5-9.1 Index]