geyser.round {locfit}	R Documentation

Discrete Old Faithful Geyser Dataset

Description

This is a variant of the geyser dataset, where each observation is rounded to the nearest 0.05 minutes, and the counts tallied.

Usage

data(geyser.round)

Format

Data Frame with variables duration and count.

Source

Scott (1992). Note that several different Old Faithful Geyser datasets (including the faithful dataset in R's base library) have been used in various places in the statistics literature. The version provided here has been used in density estimation and bandwidth selection work.

References

Scott, D. W. (1992). Multivariate Density Estimation: Theory, Practice and Visualization. Wiley.

[Package locfit version 1.5-9.1 Index]