hatmatrix {locfit}	R Documentation

Weight diagrams and the hat matrix for a local regression model.

Description

hatmatrix() computes the weight diagrams (also known as equivalent or effective kernels) for a local regression smooth. Essentially, hatmatrix() is a front-end to locfit(), setting a flag to compute and return weight diagrams, rather than the fit.

Usage

hatmatrix(formula, dc=TRUE, ...)

Arguments

formula
model formula.

dc
derivative adjustment (see locfit.raw)

...
Other arguments to locfit and locfit.raw.

Value

A matrix with n rows and p columns; each column being the weight diagram for the corresponding locfit fit point. If ev="data", this is the transpose of the hat matrix.

See Also

locfit, plot.locfit.1d, plot.locfit.2d, plot.locfit.3d, lines.locfit, predict.locfit

[Package locfit version 1.5-9.1 Index]