kappa0 {locfit}	R Documentation

Critical Values for Simultaneous Confidence Bands.

Description

The geometric constants for simultaneous confidence bands are computed, as described in Sun and Loader (1994) (bias adjustment is not implemented here). These are then passed to the crit function, which computes the critical value for the confidence bands.

The method requires both the weight diagrams l(x), the derivative l'(x) and (in 2 or more dimensions) the second derivatives l”(x). These are implemented exactly for a constant bandwidth. For nearest neighbor bandwidths, the computations are approximate and a warning is produced.

The theoretical justification for the bands uses normality of the random errors e_1,…,e_n in the regression model, and in particular the spherical symmetry of the error vector. For non-normal distributions, and likelihood models, one relies on central limit and related theorems.

Computation uses the product Simpson's rule to evaluate the multidimensional integrals (The domain of integration, and hence the region of simultaneous coverage, is determined by the flim argument). Expect the integration to be slow in more than one dimension. The mint argument controls the precision.

Usage

kappa0(formula, cov=0.95, ev=lfgrid(20), ...)

Arguments

formula
Local regression model formula. A "locfit" object can also be provided; in this case the formula and other arguments are extracted from this object.

cov
Coverage Probability for critical values.

ev
Locfit evaluation structure. Should usually be a grid – this specifies the integration rule.

...
Other arguments to locfit. Important arguments include flim and alpha.

Value

A list with components for the critical value, geometric constants, e.t.c. Can be passed directly to plot.locfit as the crit argument.

References

Sun, J. and Loader, C. (1994). Simultaneous confidence bands for linear regression and smoothing. Annals of Statistics 22, 1328-1345.

See Also

locfit, plot.locfit, crit, crit<-.

Examples

# compute and plot simultaneous confidence bands
data(ethanol)
fit <- locfit(NOx~E,data=ethanol)
crit(fit) <- kappa0(NOx~E,data=ethanol)
plot(fit,crit=crit,band="local")

[Package locfit version 1.5-9.1 Index]