kdeb {locfit}	R Documentation

Bandwidth selectors for kernel density estimation.

Description

Function to compute kernel density estimate bandwidths, as used in the simulation results in Chapter 10 of Loader (1999).

This function is included for comparative purposes only. Plug-in selectors are based on flawed logic, make unreasonable and restrictive assumptions and do not use the full power of the estimates available in Locfit. Any relation between the results produced by this function and desirable estimates are entirely coincidental.

Usage

kdeb(x, h0 = 0.01 * sd, h1 = sd, meth = c("AIC", "LCV", "LSCV", "BCV", 
  "SJPI", "GKK"), kern = "gauss", gf = 2.5)

Arguments

x
One dimensional data vector.

h0
Lower limit for bandwidth selection. Can be fairly small, but h0=0 would cause problems.

h1
Upper limit.

meth
Required selection method(s).

kern
Kernel. Most methods require kern="gauss", the default for this function only.

gf
Standard deviation for the gaussian kernel. Default 2.5, as Locfit's standard. Most papers use 1.

Value

Vector of selected bandwidths.

References

Loader, C. (1999). Local Regression and Likelihood. Springer, New York.

[Package locfit version 1.5-9.1 Index]