lcv {locfit}	R Documentation

Compute Likelihood Cross Validation Statistic.

Description

The calling sequence for lcv matches those for the locfit or locfit.raw functions. The fit is not returned; instead, the returned object contains likelihood cross validation score for the fit.

The LCV score is exact (up to numerical roundoff) if the ev="cross" argument is provided. Otherwise, the influence and cross validated residuals are computed using locfit's standard interpolation based approximations.

Usage

lcv(x, ...)

Arguments

x
model formula

...
other arguments to locfit

See Also

locfit, locfit.raw, lcvplot

[Package locfit version 1.5-9.1 Index]