lfgrid {locfit}	R Documentation

Locfit - grid evaluation structure.

Description

lfgrid() is used to specify evaluation on a grid of points for locfit.raw(). The structure computes a bounding box for the data, and divides that into a grid with specified margins.

Usage

lfgrid(mg=10, ll, ur)

Arguments

mg
Number of grid points along each margin. Can be a single number (which is applied in each dimension), or a vector specifying a value for each dimension.

ll
Lower left limits for the grid. Length should be the number of dimensions of the data provided to locfit.raw().

ur
Upper right limits for the grid. By default, ll and ur are generated as the bounding box for the data.

Examples

data(ethanol, package="locfit")
plot.eval(locfit(NOx ~ lp(E, C, scale=TRUE), data=ethanol, ev=lfgrid()))

[Package locfit version 1.5-9.1 Index]