lfknots {locfit}	R Documentation

Extraction of fit-point information from a Locfit object.

Description

Extracts information, such as fitted values, influence functions from a "locfit" object.

Usage

lfknots(x, tr, what = c("x", "coef", "h", "nlx"), delete.pv = TRUE)

Arguments

x
Fitted object from locfit().

tr
Back transformation. Default is the invers link function from the Locfit object.

what
What to return; default is c("x","coef","h","nlx"). Allowed fields are x (fit points); coef (fitted values); f1 (local slope); nlx (length of the weight diagram); nlx1 (estimated derivative of nlx); se (standard errors); infl (influence function); infla (slope of influence function); lik (maximixed local log-likelihood and local degrees of freedom); h (bandwidth) and deg (degree of fit).

delete.pv
If T, pseudo-vertices are deleted.

Value

A matrix with one row for each fit point. Columns correspond to the specified what vector; some fields contribute multiple columns.

[Package locfit version 1.5-9.1 Index]