lfmarg {locfit}	R Documentation

Generate grid margins.

Description

This function is usually called by plot.locfit.

Usage

lfmarg(xlim, m = 40)

Arguments

xlim
Vector of limits for the grid. Should be of length 2*d; the first d components represent the lower left corner, and the next d components the upper right corner. Can also be a "locfit" object.

m
Number of points for each grid margin. Can be a vector of length d.

Value

A list, whose components are the d grid margins.

See Also

locfit, plot.locfit

[Package locfit version 1.5-9.1 Index]