lines.locfit {locfit}	R Documentation

Add locfit line to existing plot

Description

Adds a Locfit line to an existing plot. llines is for use within a panel function for Lattice.

Usage

## S3 method for class 'locfit'
lines(x, m=100, tr=x$trans, ...)
## S3 method for class 'locfit'
llines(x, m=100, tr=x$trans, ...)

Arguments

x
locfit object. Should be a model with one predictor.

m
Number of points to evaluate the line at.

tr
Transformation function to use for plotting. Default is the inverse link function, or the identity function if derivatives are required.

...
Other arguments to the default lines function.

See Also

locfit, plot.locfit, lines

[Package locfit version 1.5-9.1 Index]