locfit.matrix {locfit}	R Documentation

Reconstruct a Locfit model matrix.

Description

Reconstructs the model matrix, and associated variables such as the response, prior weights and censoring indicators, from a locfit object. This is used by functions such as fitted.locfit; it is not normally called directly. The function will only work properly if the data frame has not been changed since the fit was constructed.

Usage

locfit.matrix(fit, data)

Arguments

fit
Locfit object

data
Data Frame.

Value

A list with variables x (the model matrix); y (the response); w (prior weights); sc (scales); ce (censoring indicator) and base (baseline fit).

See Also

locfit, fitted.locfit, residuals.locfit

[Package locfit version 1.5-9.1 Index]