locfit.quasi {locfit}	R Documentation

Local Quasi-Likelihood with global reweighting.

Description

locfit.quasi assumes a specified mean-variance relation, and performs iterartive reweighted local regression under this assumption. This is appropriate for local quasi-likelihood models, and is an alternative to specifying a family such as "qpoisson".

locfit.quasi is designed as a front end to locfit.raw with data vectors, or as an intemediary between locfit and locfit.raw with a model formula. If you can stand the syntax, the second calling sequence above will be slightly more efficient than the third.

Usage

locfit.quasi(x, y, weights, ..., iter=3, var=abs)

Arguments

x
Either a locfit model formula or a numeric vector of the predictor variable.

y
If x is numeric, y gives the response variable.

weights
Case weights to use in the fitting.

...
Other arguments to locfit.raw

iter
Number of EM iterations to perform

var
Function specifying the assumed relation between the mean and variance.

Value

"locfit" object.

See Also

locfit, locfit.raw

[Package locfit version 1.5-9.1 Index]