locfit.robust {locfit}	R Documentation

Robust Local Regression

Description

locfit.robust implements a robust local regression where outliers are iteratively identified and downweighted, similarly to the lowess method (Cleveland, 1979). The iterations and scale estimation are performed on a global basis.

The scale estimate is 6 times the median absolute residual, while the robust downweighting uses the bisquare function. These are performed in the S code so easily changed.

This can be interpreted as an extension of M estimation to local regression. An alternative extension (implemented in locfit via family="qrgauss") performs the iteration and scale estimation on a local basis.

Usage

locfit.robust(x, y, weights, ..., iter=3)

Arguments

x
Either a locfit model formula or a numeric vector of the predictor variable.

y
If x is numeric, y gives the response variable.

weights
weights to use in the fitting.

...
Other arguments to locfit.raw.

iter
Number of iterations to perform

Value

"locfit" object.

References

Cleveland, W. S. (1979). Robust locally weighted regression and smoothing scatterplots. J. Amer. Statist. Assn. 74, 829-836.

See Also

locfit, locfit.raw

[Package locfit version 1.5-9.1 Index]