lp {locfit}	R Documentation

Local Polynomial Model Term

Description

lp is a local polynomial model term for Locfit models. Usually, it will be the only term on the RHS of the model formula.

Smoothing parameters should be provided as arguments to lp(), rather than to locfit().

Usage

lp(..., nn, h, adpen, deg, acri, scale, style)

Arguments

...
Predictor variables for the local regression model.

nn
Nearest neighbor component of the smoothing parameter. Default value is 0.7, unless either h or adpen are provided, in which case the default is 0.

h
The constant component of the smoothing parameter. Default: 0.

adpen
Penalty parameter for adaptive fitting.

deg
Degree of polynomial to use.

acri
Criterion for adaptive bandwidth selection.

style
Style for special terms (left, ang e.t.c.). Do not try to set this directly; call locfit instead.

scale
A scale to apply to each variable. This is especially important for multivariate fitting, where variables may be measured in non-comparable units. It is also used to specify the frequency for ang terms. If scale=F (the default) no scaling is performed. If scale=T, marginal standard deviations are used. Alternatively, a numeric vector can provide scales for the individual variables.

See Also

locfit, locfit.raw

Examples

data(ethanol, package="locfit")
# fit with 50% nearest neighbor bandwidth.
fit <- locfit(NOx~lp(E,nn=0.5),data=ethanol)
# bivariate fit.
fit <- locfit(NOx~lp(E,C,scale=TRUE),data=ethanol)

# density estimation
data(geyser, package="locfit")
fit <- locfit.raw(lp(geyser,nn=0.1,h=0.8))

[Package locfit version 1.5-9.1 Index]