lscv.exact {locfit}	R Documentation

Exact LSCV Calculation

Description

This function performs the exact computation of the least squares cross validation statistic for one-dimensional kernel density estimation and a constant bandwidth.

At the time of writing, it is implemented only for the Gaussian kernel (with the standard deviation of 0.4; Locfit's standard).

Usage

lscv.exact(x, h=0)

Arguments

x
Numeric data vector.

h
The bandwidth. If x is constructed with lp(), the bandwidth should be given there instead.

Value

A vector of the LSCV statistic and the fitted degrees of freedom.

See Also

lscv, lscvplot

Examples

data(geyser, package="locfit")
lscv.exact(lp(geyser,h=0.25))
# equivalent form using lscv
lscv(lp(geyser, h=0.25), exact=TRUE)

[Package locfit version 1.5-9.1 Index]