lscvplot {locfit}	R Documentation

Compute the LSCV plot.

Description

The lscvplot function loops through calls to the lscv function (and hence to link{locfit}), using a different smoothing parameter for each call. The returned structure contains the LSCV statistic for each density estimate, and can be used to produce an LSCV plot.

Usage

lscvplot(..., alpha)

Arguments

...
arguments to the lscv, locfit functions.

alpha
Matrix of smoothing parameters. The lscvplot function loops through calls to lscv, using each row of alpha as the smoothing parameter in turn. If alpha is provided as a vector, it will be converted to a one-column matrix, thus interpreting each component as a nearest neighbor smoothing parameter.

Value

An object with class "gcvplot", containing the smoothing parameters and LSCV scores. The actual plot is produced using plot.gcvplot.

See Also

locfit, locfit.raw, gcv, lscv, plot.gcvplot

[Package locfit version 1.5-9.12 Index]