mcyc {locfit}	R Documentation

Acc(De?)celeration of a Motorcycle Hitting a Wall

Description

Measurements of the acceleration of a motorcycle as it hits a wall. Actually, rumored to be a concatenation of several such datasets.

Usage

data(mcyc)

Format

Data frame with time and accel variables.

Source

H\"ardle (1990).

References

H\"ardle, W. (1990). Applied Nonparametric Regression. Cambridge University Press.

[Package locfit version 1.5-9.1 Index]