mine {locfit}	R Documentation

Fracture Counts in Coal Mines

Description

The number of fractures in the upper seam of coal mines, and four predictor variables. This dataset can be modeled using Poisson regression.

Usage

data(mine)

Format

A dataframe with the response frac, and predictor variables extrp, time, seamh and inb.

Source

Myers (1990).

References

Myers, R. H. (1990). Classical and Modern Regression with Applications (Second edition). PWS-Kent Publishing, Boston.

[Package locfit version 1.5-9.1 Index]