none {locfit}	R Documentation

Locfit Evaluation Structure

Description

none() is an evaluation structure for locfit.raw(), specifying no evaluation points. Only the initial parametric fit is computed - this is the easiest and most efficient way to coerce Locfit into producing a parametric regression fit.

Usage

none()

Examples

data(ethanol, package="locfit")
# fit a fourth degree polynomial using locfit
fit <- locfit(NOx~E,data=ethanol,deg=4,ev=none())
plot(fit,get.data=TRUE)

[Package locfit version 1.5-9.1 Index]