plot.eval {locfit}	R Documentation

Plot evaluation points from a 2-d locfit object.

Description

This function is used to plot the evaluation structure generated by Locfit for a two dimensional fit. Vertices of the tree structure are displayed as O; pseudo-vertices as *.

Usage

plot.eval(x, add=FALSE, text=FALSE, ...)

Arguments

x
"locfit" object.

add
If TRUE, add to existing plot.

text
If TRUE, numbers will be added indicating the order points were added.

...
Arguments passed to and from other methods.

See Also

locfit.

Examples

data(ethanol, package="locfit")
fit <- locfit(NOx ~ E + C, data=ethanol, scale=0)
plot.eval(fit)

[Package locfit version 1.5-9.1 Index]