plot.lfeval {locfit}	R Documentation

Plot a Locfit Evaluation Structure.

Description

Plots the evaluation points from a locfit or lfeval structure, for one- or two-dimensional fits.

Usage

## S3 method for class 'lfeval'
plot(x, add=FALSE, txt=FALSE, ...)

Arguments

x
A lfeval or locfit object

add
If TRUE, the points will be added to the existing plot. Otherwise, a new plot is created.

txt
If TRUE, the points are annotated with numbers in the order they were entered into the fit.

...
Additional graphical parameters.

Value

"lfeval" object.

See Also

lfeval, locfit, print.lfeval

[Package locfit version 1.5-9.1 Index]