plot.locfit.1d {locfit}	R Documentation

Plot a one dimensional preplot.locfit object.

Description

This function is not usually called directly. It will be called automatically when plotting a one-dimensional locfit or preplot.locfit object.

Usage

## S3 method for class 'locfit.1d'
plot(x, add=FALSE, main="", xlab="default", ylab=x$yname,
  type="l", ylim, lty=1, col=1, ...)

Arguments

x
One dimensional preplot.locfit object.

add
If TRUE, the plot will be added to the existing plot.

main, xlab, ylab, type, ylim, lty, col
Graphical parameters passed on to plot (only if add=FALSE).

...
Additional graphical parameters to the plot function (only if add=FALSE).

See Also

locfit, plot.locfit, preplot.locfit

[Package locfit version 1.5-9.1 Index]