plot.locfit.3d {locfit}	R Documentation

Plot a high-dimensional "preplot.locfit" object using trellis displays.

Description

This function plots cross-sections of a Locfit model (usually in three or more dimensions) using trellis displays. It is not usually called directly, but is invoked by plot.locfit.

The R libraries lattice and grid provide a partial (at time of writing) implementation of trellis. Currently, this works with one panel variable.

Usage

## S3 method for class 'locfit.3d'
plot(x, main="", pv, tv, type = "level", pred.lab = x$vnames,
               resp.lab=x$yname, crit = 1.96, ...)

Arguments

x
"preplot.locfit" object.

main
title for the plot.

pv
Panel variables. These are the variables (either one or two) that are varied within each panel of the display.

tv
Trellis variables. These are varied from panel to panel of the display.

type
Type of display. When there are two panel variables, the choices are "contour", "level" and "persp".

pred.lab
label for the predictor variable.

resp.lab
label for the response variable.

crit
critical value for the confidence level.

...
graphical parameters passed to xyplot or contourplot.

See Also

plot.locfit, preplot.locfit

[Package locfit version 1.5-9.1 Index]