plot.preplot.locfit {locfit}	R Documentation

Plot a "preplot.locfit" object.

Description

The plot.locfit() function is implemented, roughly, as a call to preplot.locfit(), followed by a call to plot.locfitpred(). For most users, there will be little need to call plot.locfitpred() directly.

Usage

## S3 method for class 'preplot.locfit'
plot(x, pv, tv, ...)

Arguments

x
A preplot.locfit object, produced by preplot.locfit().

pv, tv, ...
Other arguments to plot.locfit.1d, plot.locfit.2d or plot.locfit.3d as appropriate.

See Also

locfit, plot.locfit, preplot.locfit, plot.locfit.1d, plot.locfit.2d, plot.locfit.3d.

[Package locfit version 1.5-9.1 Index]