plot.scb {locfit}	R Documentation

Plot method for simultaneous confidence bands

Description

Plot method for simultaneous confidence bands created by the scb function.

Usage

## S3 method for class 'scb'
plot(x, add=FALSE, ...)

Arguments

x
scb object created by scb.

add
If TRUE, bands will be added to the existing plot.

...
Arguments passed to and from other methods.

See Also

scb

Examples

# corrected confidence bands for a linear logistic model
data(insect)
fit <- scb(deaths ~ lconc, type=4, w=nins, data=insect,
           deg=1, family="binomial", kern="parm")
plot(fit)

[Package locfit version 1.5-9.1 Index]