predict.locfit {locfit}	R Documentation

Prediction from a Locfit object.

Description

The locfit function computes a local fit at a selected set of points (as defined by the ev argument). The predict.locfit function is used to interpolate from these points to any other points. The method is based on cubic hermite polynomial interpolation, using the estimates and local slopes at each fit point.

The motivation for this two-step procedure is computational speed. Depending on the sample size, dimension and fitting procedure, the local fitting method can be expensive, and it is desirable to keep the number of points at which the direct fit is computed to a minimum. The interpolation method used by predict.locfit() is usually much faster, and can be computed at larger numbers of points.

Usage

## S3 method for class 'locfit'
predict(object, newdata=NULL, where = "fitp",
          se.fit=FALSE, band="none", what="coef", ...)

Arguments

object
Fitted object from locfit().

newdata
Points to predict at. Can be given in several forms: vector/matrix; list, data frame.

se.fit
If TRUE, standard errors are computed along with the fitted values.

where, what, band
arguments passed on to preplot.locfit.

...
Additional arguments to preplot.locfit.

Value

If se.fit=F, a numeric vector of predictors. If se.fit=T, a list with components fit, se.fit and residual.scale.

Examples

data(ethanol, package="locfit")
fit <- locfit(NOx ~ E, data=ethanol)
predict(fit,c(0.6,0.8,1.0))

[Package locfit version 1.5-9.1 Index]