preplot.locfit {locfit}	R Documentation

Prediction from a Locfit object.

Description

preplot.locfit can be called directly, although it is more usual to call plot.locfit or predict.locfit. The advantage of preplot.locfit is in S-Plus 5, where arithmetic and transformations can be performed on the "preplot.locfit" object.

plot(preplot(fit)) is essentially synonymous with plot(fit).

Usage

## S3 method for class 'locfit'
preplot(object, newdata=NULL, where, tr=NULL, what="coef",
  band="none", get.data=FALSE, f3d=FALSE, ...)

Arguments

object
Fitted object from locfit().

newdata
Points to predict at. Can be given in several forms: vector/matrix; list, data frame.

where
An alternative to newdata. Choices include "grid" for the grid lfmarg(object); "data" for the original data points and "fitp" for the direct fitting points (ie. no interpolation).

tr
Transformation for likelihood models. Default is the inverse of the link function.

what
What to compute predicted values of. The default, what="coef", works with the fitted curve itself. Other choices include "nlx" for the length of the weight diagram; "infl" for the influence function; "band" for the bandwidth; "degr" for the local polynomial degree; "lik" for the maximized local likelihood; "rdf" for the local residual degrees of freedom and "vari" for the variance function. The interpolation algorithm for some of these quantities is questionable.

band
Compute standard errors for the fit and include confidence bands on the returned object. Default is "none". Other choices include "global" for bands using a global variance estimate; "local" for bands using a local variance estimate and "pred" for prediction bands (at present, using a global variance estimate). To obtain the global variance estimate for a fit, use rv. This can be changed with rv<-. Confidence bands, by default, are 95%, based on normal approximations and neglecting bias. To change the critical value or confidence level, or to obtain simultaneous instead of pointwise confidence, the critical value stored on the fit must be changed. See the kappa0 and crit functions.

get.data
If TRUE, the original data is attached to the returned object, and added to the plot.

f3d
If TRUE, sets a flag that forces ploting using the trellis style. Not available in R.

...
arguments passed to and from other methods.

Value

An object with class "preplot.locfit", containing the predicted values and additional information used to construct the plot.

See Also

locfit, predict.locfit, plot.locfit.

[Package locfit version 1.5-9.1 Index]