preplot.locfit.raw {locfit}	R Documentation

Prediction from a Locfit object.

Description

preplot.locfit.raw is an internal function used by predict.locfit and preplot.locfit. It should not normally be called directly.

Usage

## S3 method for class 'locfit.raw'
preplot(object, newdata, where, what, band, ...)

Arguments

object
Fitted object from locfit().

newdata
New data points.

where
Type of data provided in newdata.

what
What to compute predicted values of.

band
Compute standard errors for the fit and include confidence bands on the returned object.

...
Arguments passed to and from other methods.

Value

A list containing raw output from the internal prediction routines.

See Also

locfit, predict.locfit, preplot.locfit.

[Package locfit version 1.5-9.1 Index]