rbox {locfit}	R Documentation

Local Regression, Likelihood and Density Estimation.

Description

rbox() is used to specify a rectangular box evaluation structure for locfit.raw(). The structure begins by generating a bounding box for the data, then recursively divides the box to a desired precision.

Usage

rbox(cut=0.8, type="tree", ll, ur)

Arguments

type
If type="tree", the cells are recursively divided according to the bandwidths at each corner of the cell; see Chapter 11 of Loader (1999). If type="kdtree", the K-D tree structure used in Loess (Cleveland and Grosse, 1991) is used.

cut
Precision of the tree; a smaller value of cut results in a larger tree with more nodes being generated.

ll
Lower left corner of the initial cell. Length should be the number of dimensions of the data provided to locfit.raw().

ur
Upper right corner of the initial cell. By default, ll and ur are generated as the bounding box for the data.

References

Loader, C. (1999). Local Regression and Likelihood. Springer, New York.

Cleveland, W. and Grosse, E. (1991). Computational Methods for Local Regression. Statistics and Computing 1.

Examples

data(ethanol, package="locfit")
plot.eval(locfit(NOx~E+C,data=ethanol,scale=0,ev=rbox(cut=0.8)))
plot.eval(locfit(NOx~E+C,data=ethanol,scale=0,ev=rbox(cut=0.3)))

[Package locfit version 1.5-9.1 Index]