regband {locfit}	R Documentation

Bandwidth selectors for local regression.

Description

Function to compute local regression bandwidths for local linear regression, implemented as a front end to locfit().

This function is included for comparative purposes only. Plug-in selectors are based on flawed logic, make unreasonable and restrictive assumptions and do not use the full power of the estimates available in Locfit. Any relation between the results produced by this function and desirable estimates are entirely coincidental.

Usage

regband(formula, what = c("CP", "GCV", "GKK", "RSW"), deg=1, ...)

Arguments

formula
Model Formula (one predictor).

what
Methods to use.

deg
Degree of fit.

...
Other Locfit options.

Value

Vector of selected bandwidths.

[Package locfit version 1.5-9.1 Index]