rv {locfit}	R Documentation

Residual variance from a locfit object.

Description

As part of the locfit fitting procedure, an estimate of the residual variance is computed; the rv function extracts the variance from the "locfit" object. The estimate used is the residual sum of squares (or residual deviance, for quasi-likelihood models), divided by the residual degrees of freedom.

For likelihood (not quasi-likelihood) models, the estimate is 1.0.

Usage

rv(fit)

Arguments

fit
"locfit" object.

Value

Returns the residual variance estimate from the "locfit" object.

See Also

locfit, rv<-

Examples

data(ethanol)
fit <- locfit(NOx~E,data=ethanol)
rv(fit)

[Package locfit version 1.5-9.12 Index]