scb {locfit}	R Documentation

Simultaneous Confidence Bands

Description

scb is implemented as a front-end to locfit, to compute simultaneous confidence bands using the tube formula method and extensions, based on Sun and Loader (1994).

Usage

scb(x, ..., ev = lfgrid(20), simul = TRUE, type = 1)

Arguments

x
A numeric vector or matrix of predictors (as in locfit.raw), or a model formula (as in locfit).

...
Additional arguments to locfit.raw.

ev
The evaluation structure to use. See locfit.raw.

simul
Should the coverage be simultaneous or pointwise?

type
Type of confidence bands. type=0 computes pointwise 95% bands. type=1 computes basic simultaneous bands with no corrections. type=2,3,4 are the centered and corrected bands for parametric regression models listed in Table 3 of Sun, Loader and McCormick (2000).

Value

A list containing the evaluation points, fit, standard deviations and upper and lower confidence bounds. The class is "scb"; methods for printing and ploting are provided.

References

Sun J. and Loader, C. (1994). Simultaneous confidence bands in linear regression and smoothing. The Annals of Statistics 22, 1328-1345.

Sun, J., Loader, C. and McCormick, W. (2000). Confidence bands in generalized linear models. The Annals of Statistics 28, 429-460.

See Also

locfit, print.scb, plot.scb.

Examples

# corrected confidence bands for a linear logistic model
data(insect)
fit <- scb(deaths~lp(lconc,deg=1), type=4, w=nins,
           data=insect,family="binomial",kern="parm")
plot(fit)

[Package locfit version 1.5-9.1 Index]