summary.gcvplot {locfit}	R Documentation

Summary method for a gcvplot structure.

Description

Computes a short summary for a generalized cross-validation plot structure

Usage

## S3 method for class 'gcvplot'
summary(object, ...)

Arguments

object
A gcvplot structure produced by a call to gcvplot, cpplot e.t.c.

...
arugments to and from other methods.

Value

A matrix with two columns; one row for each fit computed in the gcvplot call. The first column is the fitted degrees of freedom; the second is the GCV or other criterion computed.

See Also

locfit, gcv, gcvplot

Examples

data(ethanol)
summary(gcvplot(NOx~E,data=ethanol,alpha=seq(0.2,1.0,by=0.05)))

[Package locfit version 1.5-9.1 Index]