trimod {locfit}	R Documentation

Generated sample from a bivariate trimodal normal mixture

Description

This is a random sample from a mixture of three bivariate standard normal components; the sample was used for the examples in Loader (1996).

Format

Data frame with 225 observations and variables x0, x1.

Source

Randomly generated in S.

References

Loader, C. R. (1996). Local Likelihood Density Estimation. Annals of Statistics 24, 1602-1618.

[Package locfit version 1.5-9.1 Index]