points.locfit {locfit}	R Documentation

Add ‘locfit’ points to existing plot

Description

This function shows the points at which the local fit was computed directly, rather than being interpolated. This can be useful if one is unsure of the validity of interpolation.

Usage

## S3 method for class 'locfit'
points(x, tr, ...)

Arguments

x
"locfit" object. Should be a model with one predictor.

tr
Back transformation.

...
Other arguments to the default points function.

See Also

locfit, plot.locfit, points

[Package locfit version 1.5-9.1 Index]