rva {locfit}	R Documentation

Substitute variance estimate on a locfit object.

Description

By default, Locfit uses the normalized residual sum of squares as the variance estimate when constructing confidence intervals. In some cases, the user may like to use alternative variance estimates; this function allows the default value to be changed.

Usage

rv(fit) <- value

Arguments

fit
"locfit" object.

value
numeric replacement value.

See Also

locfit(), rv(), plot.locfit()

[Package locfit version 1.5-9.1 Index]